+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Administrators can now define the lowest role that can create events
+### Changed
+- Roster is now a group command
+- Settings is now a group command
+
 ## [1.1.2] - 2017-09-09
 ### Fixed
 - Invalid Bungie account names would cause the register command to be unresponsive